requirements.txt
anthropic>=0.88.0
fastapi>=0.135.0
uvicorn>=0.42.0
python-dotenv>=1.0.0
